Translation for db.model... and db.datatypes...
Realised some redundancies, so I omitted some of them, and this will
IMPLICATE a lot of changes in the future, but the payoff will be that
it becomes simple to and practical to understand.

Changed how the TableField works all around... It's not a data structure
anymore, it's now just a 'metadata' — metastructure or something. It
doesn't hold data for fields anymore, but rather now describes what the
field should be.

And the TableModel was moved into the Table class which was translated
from the Table trait in PHP. Realised along the way that since the trait
became a class, it's now sound to integrate combine it to that class
since the reason of making TableModel in the first place is to convert
the trait class into an object where the arbiter files can interface
with it conveniently.