Done with query.arbiter and a bit with adit.arbiter. Move to TypeScript??
Gonna finish adit.arbiter soon. And a lot of overhauls
was made with the query.arbiter translation. But it would be
easier to work with now, more simple but still performant, i think.

Implementing new features can be started now, like the JOIN and Subquery
stuffs. For the subqueries, I suggest making a new class which returns
an object to be accepted by the entries().filter() method which parses
that that object to a suitable subquery string.

I'm just gonna head on finish adit, and propagate the changes to the
Querytype and Queryset. good luck.